Backend_protected/requirements.txt
alembic==1.14.0  # Database migration tool for SQLAlchemy
annotated-types==0.7.0  # Types for Python's type hinting
anyio==4.7.0  # A library for asynchronous I/O operations, compatible with multiple backends
colorama==0.4.6  # Cross-platform colored terminal text for Python
fastapi==0.115.6  # Fast API framework for building APIs quickly
h11==0.14.0  # HTTP/1.1 library for Python, used in ASGI apps
idna==3.10  # Internationalized Domain Names (IDN) library for Python
joblib==1.4.2  # Efficient serialization of Python objects for big data
loguru==0.7.3  # Python logging library for simpler and more flexible logging
Mako==1.3.8  # Template library for Python, used for rendering HTML, XML, etc.
MarkupSafe==3.0.2  # Library for safe handling of untrusted input in templates
pydantic==2.10.3  # Data validation and settings management library, based on Python type annotations
pydantic_core==2.27.1  # Core component of Pydantic used for validation
PyMySQL==1.1.1  # Pure-Python MySQL client for Python
python-dateutil==2.9.0.post0  # Extensions to the standard Python datetime module
python-decouple==3.8  # Configuration management for Python projects
pytz==2024.1  # World timezone definitions and conversion library
six==1.16.0  # Python 2 and 3 compatibility library
sniffio==1.3.1  # Detect which async library is running in the current process
SQLAlchemy==2.0.36  # SQL toolkit and Object-Relational Mapping (ORM) for Python
starlette==0.41.3  # ASGI framework for building web apps and microservices
threadpoolctl==3.5.0  # Control the number of threads in native libraries like NumPy
typing_extensions==4.12.2  # Backports of new type hinting features to older versions of Python
tzdata==2024.1  # Timezone data library for Python
uvicorn==0.34.0  # ASGI server for Python web apps, serving FastAPI apps
win32_setctime==1.2.0  # Windows-specific module to set file timestamps
cryptography==44.0.0  # Library for cryptographic recipes and algorithms
email-validator==2.2.0  # Email address validation library for Python
python-multipart==0.0.20  # Library to handle form data in multipart format for APIs
pydantic[email]  # Pydantic with additional support for email validation
python-dotenv==1.0.1  # Manage environment variables in Python applications
passlib==1.7.4  # Password hashing library, supporting various hashing algorithms